update, delete data

+++ app/Http/Controllers/TodoController.php
@@ -10,12 +10,19 @@ class TodoController extends Controller
     public function index()
     {
         $todos = Todo::where('user_id', auth()->user()->id)
-            ->orderBy('is_done', 'asc')
+            ->orderBy('is_complete', 'asc')
             ->orderBy('created_at', 'desc')
             ->get();
     
-        return view('todo.index', compact('todos'));
+        // return view('todo.index', compact('todos'));
+        $todosCompleted = Todo::where('user_id', auth()->user()->id)
+                      ->where('is_complete', true)
+                      ->count();
+
+        return view('todo.index', compact('todos', 'todosCompleted'));
+
     }
+
 
     public function store(Request $request, Todo $todo)
     {
@@ -35,20 +42,90 @@ class TodoController extends Controller
             ->with('success', 'Todo created successfully!');
     }
     
-    public function destroy(Todo $todo)
-    {
-        $todo->delete();
-    
-        return redirect()->route('todo.index')->with('success', 'Todo deleted successfully.');
-    }
 
     public function create()
     {
         return view('todo.create');
     }
 
-    public function edit()
-    {
-        return view('todo.edit');
+    public function edit(Todo $todo)
+{
+    if (auth()->user()->id == $todo->user_id) {
+        // dd($todo);
+        return view('todo.edit', compact('todo'));
+    } else {
+        // abort(403);
+        // abort(403, 'Not authorized');
+        return redirect()->route('todo.index')->with('danger', 'You are not authorized to edit this todo!');
     }
+}
+
+public function update(Request $request, Todo $todo)
+{
+    $request->validate([
+        'title' => 'required|max:255',
+    ]);
+
+    // Practical
+    // $todo->title = $request->title;
+    // $todo->save();
+
+    // Eloquent Way - Readable
+    $todo->update([
+        'title' => ucfirst($request->title),
+    ]);
+
+    return redirect()->route('todo.index')->with('success', 'Todo updated successfully!');
+}
+
+
+public function complete(Todo $todo)
+{
+    if (auth()->user()->id == $todo->user_id) {
+        $todo->update([
+            'is_complete' => true,
+        ]);
+        return redirect()->route('todo.index')->with('success', 'Todo completed successfully!');
+    } else {
+        return redirect()->route('todo.index')->with('danger', 'You are not authorized to complete this todo!');
+    }
+}
+
+public function uncomplete(Todo $todo)
+{
+    if (auth()->user()->id == $todo->user_id) {
+        $todo->update([
+            'is_complete' => false,
+        ]);
+        return redirect()->route('todo.index')->with('success', 'Todo uncompleted successfully!');
+    } else {
+        return redirect()->route('todo.index')->with('danger', 'You are not authorized to uncomplete this todo!');
+    }
+}
+
+public function destroy(Todo $todo)
+{
+    if (auth()->user()->id == $todo->user_id) {
+        $todo->delete();
+        return redirect()->route('todo.index')->with('success', 'Todo deleted successfully!');
+    } else {
+        return redirect()->route('todo.index')->with('danger', 'You are not authorized to delete this todo!');
+    }
+}
+
+public function destroyCompleted()
+{
+    // get all todos for current user where is_doned is true
+    $todosCompleted = Todo::where('user_id', auth()->user()->id)
+                          ->where('is_complete', true)
+                          ->get();
+
+    foreach ($todosCompleted as $todo) {
+        $todo->delete();
+    }
+
+    return redirect()->route('todo.index')->with('success', 'All completed todos deleted successfully!');
+}
+
+
 }
